requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
#Flask-JWT-Extended==4.2.3
psycopg2-binary==2.9.9
generateapikey==0.4.5
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
PyYAML==6.0.1